common: fold object_fmt.h into object.h
as fmtlib is now an accepted dependency in all of Ceph.
